backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3

# Database
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.4
redis==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# ML & NLP
scikit-learn==1.3.2

# LLM Integration
openai==1.3.7
anthropic==0.7.7

# Report Generation
reportlab==4.0.7
pypdf==3.17.0

# Utilities
python-dotenv==1.0.0
python-magic==0.4.27
hashlib-additional==1.0.0

# Auth
djangorestframework-simplejwt==5.3.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
